requirements.txt
flask==3.1.0
langchain==0.3.14
langchain_community==0.3.14
langchain-openai==0.3.4
langchain-core==0.3.45
neo4j==5.24.0
sentence_transformers==3.3.1
langchain_groq==0.2.4
langchain-google-genai==2.1.0
langchain-qdrant==0.2.0
langchain_huggingface==0.1.2
tf_keras==2.18.0
# ragatouille==0.0.8.post4

dotenv==0.9.9
python-dotenv==1.1.0
PyYAML==6.0.2
annotated-types==0.7.0
anyio==4.9.0
fastapi==0.115.12
h11==0.14.0
httptools==0.6.4
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.13.0
uvicorn==0.34.0
